+++ rs-week1/src/__tests__/Page404.test.tsx
@@ -10,7 +10,6 @@ describe('<Page404 />', () => {
     const wrapper = render(<Page404 onChangeNamePage={onChangeNamePage} />);
     expect(wrapper).toBeTruthy();
 
-    // Get by h3
     const h3 = wrapper.container.querySelector('h3');
     expect(h3?.textContent).toMatch(/404/i);
   });
