+++ rs-week1/src/components/types/enum.ts
@@ -1,0 +1,7 @@
+enum Ecategory {
+  smartphones = 'smartphones',
+  laptops = 'laptops',
+  fragrances = 'fragrances',
+}
+
+export { Ecategory };
